backend/requirements.txt
# Django & Core
Django==4.2.29
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1
drf-spectacular==0.29.0

# Database
mysqlclient==2.1.1

# Config & Utils
python-decouple==3.8
pillow==11.3.0
requests==2.32.3

# PDF & Document
PyPDF2==3.0.1
reportlab==4.4.10
python-docx==1.2.0
fpdf2==2.8.3

# Production Server
gunicorn==23.0.0

# Logging (production)
python-json-logger==3.3.0

# Error Tracking (optional)
sentry-sdk==2.22.0
